qml: move backend models into models directory
