+++ src/qml/bitcoin.cpp
@@ -13,14 +13,14 @@
 #include <node/interface_ui.h>
 #include <noui.h>
 #include <qml/appmode.h>
-#include <qml/chainmodel.h>
 #include <qml/components/blockclockdial.h>
 #include <qml/controls/linegraph.h>
-#include <qml/networktraffictower.h>
+#include <qml/models/chainmodel.h>
+#include <qml/models/networktraffictower.h>
+#include <qml/models/nodemodel.h>
+#include <qml/models/options_model.h>
+#include <qml/models/peerlistsortproxy.h>
 #include <qml/imageprovider.h>
-#include <qml/nodemodel.h>
-#include <qml/options_model.h>
-#include <qml/peerlistsortproxy.h>
 #include <qml/util.h>
 #include <qt/guiconstants.h>
 #include <qt/guiutil.h>
