+++ src/qml/models/chainmodel.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <qml/chainmodel.h>
+#include <qml/models/chainmodel.h>
 
 #include <QDateTime>
 #include <QString>
