+++ src/qml/models/networktraffictower.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <qml/networktraffictower.h>
+#include <qml/models/networktraffictower.h>
 
 #include <QObject>
 #include <QThread>
