+++ src/qml/models/nodemodel.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <qml/nodemodel.h>
+#include <qml/models/nodemodel.h>
 
 #include <interfaces/node.h>
 #include <net.h>
