+++ src/qml/models/options_model.h
@@ -2,8 +2,8 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#ifndef BITCOIN_QML_OPTIONS_MODEL_H
-#define BITCOIN_QML_OPTIONS_MODEL_H
+#ifndef BITCOIN_QML_MODELS_OPTIONS_MODEL_H
+#define BITCOIN_QML_MODELS_OPTIONS_MODEL_H
 
 #include <txdb.h>
 #include <util/settings.h>
@@ -87,4 +87,4 @@ private:
     util::SettingsValue pruneSetting() const;
 };
 
-#endif // BITCOIN_QML_OPTIONS_MODEL_H
+#endif // BITCOIN_QML_MODELS_OPTIONS_MODEL_H
