+++ src/qml/models/peerlistsortproxy.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <qml/peerlistsortproxy.h>
+#include <qml/models/peerlistsortproxy.h>
 #include <qt/peertablemodel.h>
 
 PeerListSortProxy::PeerListSortProxy(QObject* parent)
